+++ src/Zope2/Startup/zopectl.py
@@ -229,12 +229,6 @@ class ZopeCmd(ZDCmd):
             args = [opt, svalue]
         return args
 
-    def do_start(self, arg):
-        # signal to Zope that it is being managed
-        # (to indicate it's web-restartable)
-        os.putenv('ZMANAGED', '1')
-        ZDCmd.do_start(self, arg)
-
     ## START OF WINDOWS ONLY STUFF
     
     if WIN: